Job near to completely functional
You can now hire/fire the NPCs, and there are some cool graphics changes
along with that, too!

Do need to add a visual aid to show if an NPC already has as job though.

Also added the working buildings to the path list, but that's just the
early work.