Added the "CraftingMenu" skeleton, as well as "Item" class
Going to try and use a "Queue" for the items that I'll be crafting.
Seems to make sense, and I get to learn something new!

Since there are going to be a ton of items, and options, it might be an
idea to start looking at doing a toolbar'd window..