Quite a bit
No need having the player anymore. It was just slow and annoying.

The camera now zooms, and has a better translation.

Changed the mine so there aren't "door locations" in the corners.

Hovering over a building shows the inside.

Adding the layers enum.

Starting to implement the buttons for the buildings.